backend/requirements.txt
django
requests
psycopg2
black
isort
django_otp
qrcode
Pillow
pyotp
python-logstash
asgiref
sqlparse
channels
channels_redis
daphne
djangorestframework
django-cors-headers
web3
pyjwt
